requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
mss>=6.0.0
pygame>=2.0.0
Pillow>=8.0.0